docs+refactor(rusmppc): added docs to test module and moved function `raw` to connection.rs
