+++ rusmppc/src/tests.rs
@@ -1,3 +1,9 @@
+//! Tests in this module test the library's functionality based on the public API.
+//!
+//! They simulate real scenarios by creating in-memory connections between a test server and a client built using the library's public API.
+//!
+//! For more in depth tests, see `connection/tests.rs`.
+
 use std::time::{Duration, Instant};
 
 use futures::{SinkExt, StreamExt};
@@ -652,7 +658,7 @@ async fn server_sends_an_operation_with_the_same_sequence_number_of_a_pending_re
     let _ = events.await;
 }
 
-/// See `server_ddos_client_should_still_send_requests_and_connection_should_still_manage_timeouts` in `connection.rs`` for a more reliable test of the same behavior.
+/// See `server_ddos_client_should_still_send_requests_and_connection_should_still_manage_timeouts` in `connection/tests.rs`` for a more reliable test of the same behavior.
 #[tokio::test]
 async fn server_ddos_client_should_still_send_requests_and_connection_should_still_manage_timeouts()
 {
